chapter-17-automation/requirements.txt
-e openai>=1.0
python-dotenv>=1.0
schedule>=1.0
